just respond with a pointer to our new website
